requirements.txt
flask
flask-sqlalchemy
flask-login
flask-mail
mysql-connector-python
pandas
openpyxl
fpdf2
werkzeug
email_validator
flask-socketio
pyjwt
python-dotenv
flask-wtf
flask-limiter
bleach
eventlet
flask-talisman
